A significant mistake was fixed in ham1_t. Several methods in the class were made more error-safe.
The mistake was in the reciprocal lattice vectors.

The major improvement is in the diag() method, which now carefully handles potential 0/0 cases, which I call "marginal". It also reports when such
cases happen using a returned bool variable. The methods that use diag() now count and report the number of marginal cases.

Furthermore, the bisec1 method (and the bisec function in the module math) now check that the return 'image' is a valid number.