Big revisions made to driver_ham1.cc: now, the parameters to be specified are specified in the constructor. Default temperature changed in ham1.h.
In order to avoid different classes for every parameter space, we now have the user specify the parameter space to be studied in the class
constructor. This is done by defining coordinate arrays for all the variable parameters. The user can specify the length of the dimensions (including
length 1) and the values they should span.

To do: Make the NetCDF class save length-1 dimensions as attributes instead of actual dimensions. Also, in pspace_t::pstudy, think of using one
variable to loop over the parameter space; this would make future changes simpler to implement.