The module ncio was modified to allow for "pruning" of dimensions of length 1. Changes not yet implemented for "manual" coordinate variables.
In effect, this is straightforward because variables are saved by passing a 1D array stored in row-major storage, so dimensions of length 1 do not
affect the placement of the data. If pruning is activated, dimensions of length 1 are skipped during the definition of dimensions. An array of dimids
that ignores these length-one dimensions is built to be used in the definition of variables. Coordinate variables for such dimensions are in fact not
saved as coordinate variables, but as "scalar" variables not associated with any dimensions.