Several changes made to the data analysis scripts.
The data_anal.py script was further modularized to allow for input of variable dictionnaries from any source. This way, it can be used to plot data
postprocessed in a script.

The nc_IO.py script was modified so that it prints the dimensions of the dataset after, rather than before, it (optionally) gets rid of the complex
dimension.

The plotting routines in data/plot_routines.py were modified to deal with complex data. This changes are tentative, as I haven't come up with a
satisfactory design for this.