The file driver_ham1.cc was rewritten using a base class and derived classes.
A variable number of dimensions is passed to the base class at construction (specifically, the number of dimensions, their names, their lengths, and
the values their coordinate variables should span). In the method pspacebase_t::pstudy(), indexing is now done with one global index. Values for the
parameters t_, tp_, J_, and x_ are calculated as a function of this index in the method parvals(), to be implemented in the derived classes.

The basic derived class is pspace_t, which just goes over a grid in parameter space.

Another derived class (**not yet tested**) is pspace_time_t, which will solve the problem for tp_ varying sinusoidally as a function of time.